+++ modules/utils/src/channel.ts
@@ -41,3 +41,11 @@ export const getBalanceForAssetId = (
   }
   return channel.balances[assetIdx].amount[participant === "alice" ? 0 : 1];
 };
+
+export const getParticipant = (channel: FullChannelState, publicIdentifier: string): "alice" | "bob" | undefined => {
+  const iAmAlice = publicIdentifier.toLowerCase() === channel.aliceIdentifier.toLowerCase();
+  if (!iAmAlice && publicIdentifier.toLowerCase() !== channel.bobIdentifier.toLowerCase()) {
+    return undefined;
+  }
+  return iAmAlice ? "alice" : "bob";
+};