+++ modules/documentation/docs/changelog.md
@@ -2,10 +2,14 @@
 
 ## Next Release
 
+- \[utils\] Create util for determining which participant
+- \[all\] Don't use pino-pretty in prod mode
+
 ## 0.1.15-beta.1
 
 - \[router\] Do not forward if deposit balance is not present in channel
 - \[router\] Fix start script to pass mnemonic secret into router
+
 ## 0.1.15-beta.0
 
 - \[utils\] Don't use `instanceof` operator